+++ src/__tests__/mock/Queen.ts
@@ -1,8 +1,0 @@
-import { injectable } from '../../injectable.decorator';
-import { KindomKing } from './King';
-
-@injectable(['KindomKing'])
-export class KindomQueen {
-    readonly name = `I'm a queen`;
-    constructor(public readonly king: KindomKing) {}
-}