+++ src/index.ts
@@ -1,4 +1,5 @@
 export { inject } from './inject.decorator';
+export { injectAll } from './injectAll.decorator';
 export { ProxyDiContainer } from './ProxyDiContainer';
 export { injectable } from './injectable.decorator';
 export { middleware } from './middleware/middleware.decorator';
@@ -14,6 +15,8 @@ export {
     DependencyClass,
     ContainerSettings,
     OnContainerizedMethod,
+    ContainerizedDependency,
+    ResolveScope,
     // TODO: I'm not sure if these should be shared
     PROXYDI_CONTAINER,
     DEPENDENCY_ID,